requirements.txt
numpy
timm
torch
torchvision
tqdm
pillow
scikit-learn
configargparse
python-dotenv
diffusers
transformers
accelerate
safetensors
sentencepiece
ftfy